requirements.txt
tkinter
pillow
keyboard